testing ll double

+++ 1-100/059/B/b1.cpp
@@ -6,7 +6,7 @@ using ll = long long ;
 
 int main()
 {
-    double A,B;
+    ll A,B;
     cin >> A >> B;
 
     if(A>B){
